Move tests
I'm about to adapt them to be based on `HalfEdge`, therefore the new
location.
